requirements/base.txt
# Django Core
Django==5.2.8
djangorestframework==3.15.2
django-environ==0.11.2
django-cors-headers==4.6.0

# Database
psycopg[binary]==3.2.12

# Authentication
djangorestframework-simplejwt==5.5.1

# API Documentation
drf-spectacular==0.28.0

# Celery
celery[redis]==5.4.0
redis==5.2.1
django-celery-beat==2.8.0
django-celery-results==2.6.0

# Monitoring & Logging
structlog==24.4.0

# Utils
python-dateutil==2.9.0
pytz==2024.2

